+++ includes/minishell.h
@@ -6,9 +6,12 @@
 /*   By: nnishiya <nnishiya@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/12 17:12:13 by nnishiya          #+#    #+#             */
-/*   Updated: 2025/09/15 12:59:08 by nnishiya         ###   ########.fr       */
+/*   Updated: 2025/09/18 15:24:41 by nnishiya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#ifndef MINI_SHELL
+# define MINI_SHELL
 
 #include <unistd.h>
 #include <stdio.h>
@@ -22,3 +25,11 @@
 #include "lexer.h"
 #include "parser.h"
 #include "libft.h"
+#include "env_table.h"
+
+int  get_exit_status(void);
+void set_exit_status(int status);
+void reset_exit_status(void);
+void inc_exit_status(int delta);
+
+#endif
