+++ src/env_table/env_table_destroy.c
@@ -1,37 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   env_table_builtin_utils.c                          :+:      :+:    :+:   */
+/*   env_table_destroy.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nnishiya <nnishiya@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/09/15 16:04:55 by nnishiya          #+#    #+#             */
-/*   Updated: 2025/09/18 10:48:12 by nnishiya         ###   ########.fr       */
+/*   Created: 2025/09/18 15:03:25 by nnishiya          #+#    #+#             */
+/*   Updated: 2025/09/18 15:03:35 by nnishiya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void print_cb(const char *k, const char *v, void *ud)
-{
-    (void)ud;
-
-    if (v) {
-        printf("%s=%s\n", k, v);
-    } else {
-        printf("%s=\n", k);
-    }
-}
-
-void env_table_foreach(const t_env_table *t, t_env_iter_cb cb, void *ud)
+void destroy_env_table(t_env_table *t)
 {
     size_t i;
 
+    if (!t || !t->pair)
+        return;
     i = 0;
     while (i < t->cap)
     {
-        if (t->pair[i].key && !t->pair[i].tomb)
-            cb(t->pair[i].key, t->pair[i].val, ud);
+        free_entry(&t->pair[i]);
         i++;
     }
+    free(t->pair);
+    t->pair = NULL;
+    t->cap = 0;
+    t->size = 0;
 }
