+++ src/env_table/utils/env_table_core.c
