+++ src/input/utils/signal.c
