+++ src/input/utils/syntax_check.c
