+++ src/input/utils/syntax_check_utils.c
